Real Mode Emulator / VBE (INCOMPLETE)
Started porting a Real Mode Emulator to C# for MOSA.
With this should come the ability to use BIOS interrupts in pmode.
First use will be for VBE purposes (to switch video modes and stuff).
It has yet to be completed, a long way to go and lots of code to port!
The original Real Mode Emulator project can be found here:
http://git.mutabah.net/rme2.git