Worked on #123 (Complete)
- SZArray now have the generic interfaces.
- Fixed a bug in Generic Instance signatures in which two signatures
would be different but the name would be the same causing crashes.
- Due to the way two MosaMethods are compared, it's best to use full
name for storage instead of the object otherwise there are conflicts
even though they are two different methods. (For MosaTypeLayout).
- MosaProperty now lazy loads the getter and setter methods as the may
change due to inheritence, etc, and we need to appropriately adjust,
this was the easiest way.
- Added a simple test for the SZArray Generic interfaces (verified using
IL DASM).